internal/urlhelper: rename package
