+++ internal/urlhelper/url_test.go
@@ -1,4 +1,4 @@
-package helper
+package urlhelper
 
 import (
 	"net/http"
@@ -80,7 +80,7 @@ func TestGetScheme(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			if got := GetScheme(tt.args.r); got != tt.want {
+			if got := getScheme(tt.args.r); got != tt.want {
 				t.Errorf("GetScheme() = %v, want %v", got, tt.want)
 			}
 		})
